[executor] refactor: split executor trait implementations to separate files
So that it's easy introduce generic implementations to them in steps.
